ssid based reporting

+++ src/wislabs.wifi.manager/dao/dao.go
@@ -170,7 +170,7 @@ type DashboardMetric struct {
 }
 
 type DashboardAppInfo struct {
-	AppId          int                   `db:"appid"json:"appid"`
+	AppId          int64                   `db:"appid"json:"appid"`
 	TenantId       int                   `db:"tenantid"json:"tenantid"`
 	Aggregate      string                `db:"aggregate"json:"aggregate"`
 	Name           string                `db:"name"json:"name"`
